+++ conin/common/conin/__init__.py
@@ -2,4 +2,4 @@ from .load_model import load_model
 from .save_model import save_model
 from .log_potential import log_potential
 from .is_polytree import is_polytree
-from .to_pgmpy import convert_conin_to_pgmpy_bn
+from .to_pgmpy_bn import convert_conin_to_pgmpy_bn
